refactor(react-gui): give the gradient section its own view domain
MultiGradSection was 668 lines, and the largest thing in it was not the
editing -- it was deciding what range of values the strip is looking at.

That range is deliberately not the stop range. A user who zooms in to
place a stop precisely must not have the view snap back every time the
stops move, so the view is an override on top of a fit baseline; and the
baseline prefers the histogram's central 95% over the raw map min/max,
because one outlier voxel otherwise compresses every stop into a sliver
at one end. During a drag the domain is frozen outright: the stops are
moving, and letting the axis move with them would make the marker appear
to stand still.

Those three rules are the hook, and they are why it is worth having a
name. What is left in the section is the editing: which stop is selected,
what a drag writes, what the toolbar does.

The stop shape goes with it. A stop carries two colours -- `color` is the
CueMol string the renderer stores, which may be a name; `hex` is what the
swatch draws -- and they diverge only between an optimistic edit and the
refetch that replaces it, which is the whole reason `resolveDisplayHex`
exists.

One more exhaustive-deps omission, the same shape as the last five
splits: a commit compares the new stop range against the frozen view and
may move the view to follow, and both of those became arguments.
